# Models/Dto/Account/UpdateUserDto.cs
namespace Beachcombing_API.Models.Dto.Account
{
    public class UpdateUserDto
    {
        public string UserName { get; set; }
        public string Email { get; set; }
        public string SelfIntro { get; set; }
       
    }
}